Make read and read_array assert correct alignment.
Unaligned accesses are UB in LLVM (and silently corrupt data on at least
some ARM chips), so before doing unsafe transmutes on the memory, we
need to ensure the incoming slice is aligned correctly for the output
type.

Add some tests to ensure this happens ... and assume the u8 arrays are
32bit aligned, so we need to drop the first element to get an unaligned
u32.  This may not actually be true (and if so, the test will be flaky)
but the test passes for several local runs on x86-64.